==> notebooks_exploration_cleaning/2_normalization.ipynb <==
# Normalization and more pre-processing of the data

In this notebook some more steps are followed to pre-process the data and normalize it. The data set in the _data/df_preproceesed.csv_ file has been obtained by merging together all the data sets of the different months of the trial, built following the code of the _data_preprocessing.ipynb_ notebook (i.e. appending the data sets _January.csv_, _February.csv_..... one under the other)

import pandas as pd
import os
import glob
import numpy as np

# 1. Cleaning and some exploratory analysis 

# Import the data
df = pd.read_csv('df_preprocessed.csv')
df

## Some generic cleaning and data exploration

df[df['speech'].isnull()] # check na

# Eliminate lines containing description about the trial used for preprocessing (manually detected)
match = ['#APPEAR', '#LOS ANGELES,', '#DEPARTMENT NO', 'ALL THE MEMB', 'ALL', 'DETECTIVE', '#LADIES', '0']
s = df.loc[df['person'].str.startswith(tuple(match)), 'person']
idx_todrop = s.index.tolist()
df.drop(df.index[idx_todrop], inplace = True)
df = df.reset_index(drop=True)

# Check the dimension of the data set
df.shape

# Substitute some errors in the transcripts (manually detected)
df[df['person'].str.contains('GOOD')]
nrow = df.shape[0]
for idx in range(0, nrow): 
    string = df.iloc[idx, 0]
    if 'GOOD' in string: 
        #df.iloc[idx, 0] = '#MR. SHAPIRO: '
        df['person'] = df['person'].replace(['MR. SHAPIRO:GOOD AFTERNOON, LADIES AND GENTLEMEN OF THE JURY.\n'],'#MR SHAPIRO: ')
    elif '402' in string:
        df['person'] = df['person'].replace([string],'[LUPER]')      

# correct some details in the person column to make the text homogeneous
df["person"].replace({
    'MS. CLARK:\n': '#MS. CLARK: ',
    '##MS. CLARK: ': '#MS. CLARK: ',
    '##MR. BLASIER: ': '#MR. BLASIER: ',
    'MR. COCHRAN:\n': '#MR. COCHRAN: ',
    'MS. LEWIS:\n' : '#MS. LEWIS: ',
    '#DET. FUHRMAN: ': "['FUHRMAN,']",
    'MS. COCHRAN:\n': '#MR. COCHRAN: ',
    'MR. SCHECK:\n':'#MR. SCHECK: ',
    '##MR. KELBERG: ': '#MR. KELBERG: ',
    '##MR. CLARKE: ': '#MR. CLARKE: ',
    '##MR. NEUFELD: ': '#MR. NEUFELD: ',
    '##MR. HARMON: ': '#MR. HARMON: ',
    'MR. GOLDBERG:\n':'#MR. GOLDBERG: ',
    'MR. GORDON:\n':'#MR. GORDON: ', 
    'MR. DARDEN:\n':'#MR. DARDEN: ',
    'MR. DOUGLAS:\n':'#MR. DOUGLAS: ',
    'MR. BAILEY:\n':'#MR. BAILEY: ',
    'MR. SHAPIRO:\n':'#MR. SHAPIRO: ', 
    'MS. CHAPMAN:\n':'#MS. CHAPMAN: '}, inplace=True)

# Exploratory analysis looking at word counts by person
counts = df.groupby('person').count().sort_values(['speech'], ascending=False)
counts

Define some lists containing relevant groups of people. Except for the _witnesses_ list, the names have been defined manually by consulting sources about the trial: 

- the __dream team__ : attorneys defending OJ Simpson
- the __prosecution__ : attorneys prosecuting OJ Simpson
- the __experts__: experts called by the dream team and prosecution to give opinions about some aspects of the trial
- the __witnesses__: witnesses of the trial. Their name is between square brackets, because of how the text has been pre-processed

dream_team = ['SHAPIRO','COCHRAN','BAILEY', 'DERSHOWITZ', 'KARDASHIAN', 'HOLLEY', 'DOUGLAS', 'UELMEN', 'SCHECK', 'NEUFELD', 'BLASIER', 'THOMPSON', 'CHAPMAN', 'CAPLAN']
prosecution = ['CLARK', 'HODGMAN', 'DARDEN', 'KELBERG', 'HARMON', 'LEWIS', 'GORDON', 'BODIN', 'GOLDBERG', 'YOCHELSON', 'DARREL' , 'LYNCH']
experts = ['CLARKE', 'DR. LAKSHMANAN', 'MR. SIMS', 'MATHESON', 'MAZZOLA', 'DR. GERDES', 'DR. COTTON', 'DEEDRICK', 'BROCKBANK', 'LEE', 'DR. WEIR']
witness_match = ['[']

# Count of speech by people group
defense_count = list()
prosecution_count = list()
experts_count = list()
nrow_counts = counts.shape[0]
witness_count = list()
for row in range(0, nrow_counts):
    string = counts.index[row]
    if any(ext in string for ext in dream_team):
        defense_count.append(counts.iloc[row, 0])
    elif any(ext in string for ext in prosecution):
        prosecution_count.append(counts.iloc[row, 0])
    elif any(ext in string for ext in experts):
        experts_count.append(counts.iloc[row, 0])
    elif any(ext in string for ext in witness_match):
        witness_count.append(counts.iloc[row, 0])
defense_volume = sum(defense_count)
prosecution_volume = sum(prosecution_count)
experts_volume = sum(experts_count)
witness_volume = sum(witness_count)
count_grouped = pd.DataFrame({'defense': [defense_volume], 'prosecution': [prosecution_volume], 'experts': [experts_volume], 'witnesses':[witness_volume]})
count_grouped

# Eliminate \n
df['speech'] = df['speech'].str.rstrip("\n")

# Add a column containing the count of words per document (document = speech)
df['number_of_words'] = df.speech.apply(lambda x: len(str(x).split()))

# Eliminate documents containing non relevant and short strings:
# i.e. documents containing greetings and having a low words count

df.drop(df[(df['speech'].str.contains('GOOD MORNING')) & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop=True)
df.drop(df[(df['speech'].str.contains('GOOD AFTERNOON')) & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop=True)
df.drop(df[(df['speech'].str.contains('GOOD EVENING')) & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop=True)
df.drop(df[(df['speech'].str.contains('YES|NO|CORRECT'))  & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop = True)
df.drop(df[(df['speech'].str.contains('CORRECT'))  & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop = True)
df.drop(df[(df['speech'].str.contains('ALL RIGHT|ALRIGHT'))  & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop = True)
df

# Eliminate part of text used for preprocessing (see ```preprocessing_transctipts_text()``` function), not relevant to text mining 
# This step might take a while

match_1 = 'CALLED AS A WITNESS BY'
match_2 = 'THE WITNESS ON THE STAND AT THE TIME'
match_3 = 'HAVING BEEN PREVIOUSLY SWORN'
match_4 = 'CROSS-EXAMINATIONBY'
match_5 = 'DIRECT EXAMINATIONBY'
match_4b = 'RECROSS-EXAMINATIONBY'
match_5b = 'REDIRECT EXAMINATIONBY'
match_6 = 'CROSS-EXAMINATION (RESUMED)BY'
match_7 = 'DIRECT EXAMINATION (RESUMED)BY'
match_6b = 'RECROSS-EXAMINATION (RESUMED)BY'
match_7b = 'REDIRECT EXAMINATION (RESUMED)BY'
nrows = df.shape[0]
for row in range(0, nrows):
    if all(match in str(df.iloc[row, 1]) for match in match_2):
        new = df.iloc[row, 1].partition("CALLED")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_2):
        new = df.iloc[row, 1].partition("WITNESS")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_3):
        new = df.iloc[row, 1].partition("HAVING")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_4):
        new = df.iloc[row, 1].partition("CROSS")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_5):
        new = df.iloc[row, 1].partition("DIRECT")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_4b):
        new = df.iloc[row, 1].partition("RECROSS")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_5b):
        new = df.iloc[row, 1].partition("REDIRECT")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_6):
        new = df.iloc[row, 1].partition("CROSS")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_7):
        new = df.iloc[row, 1].partition("DIRECT")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_6b):
        new = df.iloc[row, 1].partition("RECROSS")[0]
        df.iloc[row, 1] = new
    elif all(match in str(df.iloc[row, 1]) for match in match_7b):
        new = df.iloc[row, 1].partition("REDIRECT")[0]
        df.iloc[row, 1] = new

## Generate a new data frame with the above normalization and pre-processing

df.to_csv('df_normalized.csv', index = False) # Generate df for next tasks

df.head()
