==> notebooks_exploration_cleaning/3 anaphora_resolution_and_negation_handling.ipynb <==
# Anaphora resolution and Negation Handling

In this notebook, we keep pre-processing the data. The main steps are: 

- further __normalization__ : set unique identifiers for the main characters of the trial and their prefix
- __negation handling__ : try to identify negation with a basic tagging method
- __anaphora resolution__ : manually solve anaphoras and references in the text

For more details about the functions used, refer to the https://github.com/MiriamGiuliani/OJSimpson-text-mining-trail-transcripts/blob/27c02196bc9ab702c2eac013132ff3b976689beb/py_scripts/py_functions.py file

import pandas as pd
import os
import glob
import numpy as np
import nltk
import re
import matplotlib
from matplotlib import pyplot as plt
from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator
from nltk.corpus import stopwords
nltk.download('stopwords')
from nb_exploration_cleaning import normalize_person_col, person_prefix_normalize, negation_handling, manual_anaphora_norm_dialogs, anaphora_norm_family_members

df = pd.read_csv('df_normalized.csv') # check na
df.isna().sum()

df = df[df['speech'].notna()]
df = df.reset_index(drop=True)
df.isna().sum()

# Fill Na (for some reason one date was missing)
nas = list(df[df.isnull().any(axis=1)].index)
for row in range(nas[0], nas[-1]+1):
    df.loc[row, 'date'] = 'MARCH 15, 1995'

# Eliminate non relevant and short strings
df.drop(df[df['speech'].str.contains('\(WITNESS COMPLIES.\)', na = False)].index, inplace = True)
df = df.reset_index(drop=True)
df.drop(df[(df["speech"].str.contains("HEARSAY|OBJECTION")==False) & (df['number_of_words']<2)].index, inplace = True)
df = df.reset_index(drop = True)
df.drop(df[(df["speech"].str.contains("RECESS")) & (df['number_of_words']<6)].index, inplace = True)
df = df.reset_index(drop = True)

df

dream_team = ['SHAPIRO','COCHRAN','BAILEY', 'DERSHOWITZ', 'KARDASHIAN', 'HOLLEY', 'DOUGLAS', 'UELMEN', 'SCHECK', 'NEUFELD', 'BLASIER', 'THOMPSON', 'CHAPMAN', 'CAPLAN']
prosecution = ['CLARK', 'HODGMAN', 'DARDEN', 'KELBERG', 'HARMON', 'LEWIS', 'GORDON', 'BODIN', 'GOLDBERG', 'YOCHELSON', 'DARREL' , 'LYNCH']
experts = ['CLARKE', 'DR. LAKSHMANAN', 'MR. SIMS', 'MATHESON', 'MAZZOLA', 'DR. GERDES', 'DR. COTTON', 'DEEDRICK', 'BROCKBANK', 'LEE', 'DR. WEIR', 'RUBIN','BODZIAK','RIEDERS', 'SPEED','MARTZ', 'MACDONELL','BADEN']
witness_match = ['[']
court = ['COURT']

# Normalization

df['norm_speech']  = df['speech'].apply(lambda x: normalize_person_col(x))

df['person'] = df['person'].apply(lambda x: person_prefix_normalize(x))

df

# Negation Handling

df['negation_handling']  = df['norm_speech'].apply(lambda x: negation_handling(x))

df

# Manual Anaphora Resolution

# Create new column to be filled
df['anaphora_solved'] = 0

# Manually normalize text based on sequence of dialogs (for more details see py_functions.py)
# This step might take a while to complete
df = manual_anaphora_norm_dialogs(df, dream_team, prosecution, court)
df

# Manually normalize text based on family relationships, based on the time in which different
# witnesses have been interrogated (for more details see py_functions.py)
df = anaphora_norm_family_members(df)

# If there are still cases where the column "anaphora_solved" is not filled, fill it with the "negation_handling" column
an_zeros = list(df[df['anaphora_solved'] == 0].index)
df.iloc[an_zeros, 7] = df.iloc[an_zeros, 6]
df

# Write CSV
df = df.to_csv('final_data_set.csv')
